frontend/requirements.txt
streamlit
streamlit-authenticator
itables
requests # To call your FastAPI
pandas
plotly
altair
pyyaml